refactor: Split out client state from ChangePassword
